+++ src/Serializer/Score/ScoreSerializer.php
@@ -25,22 +25,42 @@ namespace OAT\Library\Lti1p3Ags\Serializer\Score;
 use OAT\Library\Lti1p3Ags\Factory\Score\ScoreFactory;
 use OAT\Library\Lti1p3Ags\Factory\Score\ScoreFactoryInterface;
 use OAT\Library\Lti1p3Ags\Model\Score\ScoreInterface;
+use OAT\Library\Lti1p3Ags\Serializer\JsonSerializer;
+use OAT\Library\Lti1p3Ags\Serializer\JsonSerializerInterface;
 use OAT\Library\Lti1p3Core\Exception\LtiException;
 use OAT\Library\Lti1p3Core\Exception\LtiExceptionInterface;
+use RuntimeException;
 
 class ScoreSerializer implements ScoreSerializerInterface
 {
     /** @var ScoreFactoryInterface */
-    private $factory;
+    private $scoreFactory;
 
-    public function __construct(?ScoreFactoryInterface $factory = null)
-    {
-        $this->factory = $factory ?? new ScoreFactory();
+    /** @var JsonSerializerInterface */
+    private $jsonSerializer;
+
+    public function __construct(
+        ?ScoreFactoryInterface $scoreFactory = null,
+        ?JsonSerializerInterface $jsonSerializer = null
+    ) {
+        $this->scoreFactory = $scoreFactory ?? new ScoreFactory();
+        $this->jsonSerializer = $jsonSerializer ?? new JsonSerializer();
     }
 
+    /**
+     * @throws LtiExceptionInterface
+     */
     public function serialize(ScoreInterface $score): string
     {
-        return json_encode($score);
+        try {
+            return $this->jsonSerializer->serialize($score);
+        } catch (RuntimeException $exception) {
+            throw new LtiException(
+                sprintf('Error during score serialization: %s', $exception->getMessage()),
+                $exception->getCode(),
+                $exception
+            );
+        }
     }
 
     /**
@@ -48,14 +68,14 @@ class ScoreSerializer implements ScoreSerializerInterface
      */
     public function deserialize(string $data): ScoreInterface
     {
-        $data = json_decode($data, true);
-
-        if (JSON_ERROR_NONE !== json_last_error()) {
+        try {
+            return $this->scoreFactory->create($this->jsonSerializer->deserialize($data));
+        } catch (RuntimeException $exception) {
             throw new LtiException(
-                sprintf('Error during score deserialization: %s', json_last_error_msg())
+                sprintf('Error during score deserialization: %s', $exception->getMessage()),
+                $exception->getCode(),
+                $exception
             );
         }
-
-        return $this->factory->create($data);
     }
 }